add: bundle file
