+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
-dist
 !.vscode
 .DS_Store
 yarn-error.log
